requirements.txt
pandas
openpyxl
firebird-driver
python-dotenv
selenium
beautifulsoup4
Faker